Save a refreshed catalog without modifying the installed package

Save a refreshed catalog without modifying the installed package

Usage

refreshDOTBasemaps(file, item_ids = NULL, overwrite = FALSE, timeout = 30)

Arguments

file

File.

item_ids

IDs.

overwrite

TRUE/FALSE.

timeout

Timeout.